Move IPC methods to their own subdirectory
* Also remove the "ipc-" prefix from them
* Update Dockerfile, compose files and package.json accordingly
