+++ ipc-methods/http-express-axios/docker-compose.yml
@@ -3,11 +3,11 @@ version: '3'
 services:
   main_process:
     build:
-      context: ../
+      context: ../../
       dockerfile: ./shared/Dockerfile
     volumes:
-      - ../results/:/usr/src/app/results
-    working_dir: /usr/src/app/ipc-http-express-axios
+      - ../../results/:/usr/src/app/results
+    working_dir: /usr/src/app/ipc-methods/http-express-axios
     environment:
       - THE_OPERATION_ENDPOINT_HOSTNAME=the_operation
       - THE_OPERATION_ENDPOINT_PORT=3000
@@ -17,9 +17,9 @@ services:
 
   the_operation:
     build:
-      context: ../
+      context: ../../
       dockerfile: ./shared/Dockerfile
-    working_dir: /usr/src/app/ipc-http-express-axios
+    working_dir: /usr/src/app/ipc-methods/http-express-axios
     environment:
       - NODE_ENV=production
     command: 'node --max-old-space-size=1536 ./server.js'
