+++ ipc-methods/http-express-axios/main.ts
@@ -1,5 +1,5 @@
 import { EIPCMethod } from 'ipc-benchmark-testing-types';
-import { mainProcess } from '../shared/mainProcess';
+import { mainProcess } from '../../shared/mainProcess';
 import { httpExpressAxiosTransportMethod } from './client';
 
 mainProcess(httpExpressAxiosTransportMethod, EIPCMethod.HTTP_EXPRESS_AXIOS);
