+++ ipc-methods/http/main.ts
@@ -1,5 +1,5 @@
 import { EIPCMethod } from 'ipc-benchmark-testing-types';
-import { mainProcess } from '../shared/mainProcess';
+import { mainProcess } from '../../shared/mainProcess';
 import { httpTransportMethod } from './client';
 
 mainProcess(httpTransportMethod, EIPCMethod.HTTP);
