+++ package.json
@@ -7,15 +7,15 @@
     "tsc": "tsc",
     "sync-data": "mkdir -p visualization-ui/src/data && rm -f visualization-ui/src/data/* && cp results/*.publish.json visualization-ui/src/data",
     "full-test-suite": "bash ./run-test-suite.sh",
-    "test:benchmark": "docker compose -f ./benchmark/docker-compose.yml up",
-    "test:benchmark:raw": "npm run tsc && node --max-old-space-size=1536 ./benchmark/main.js",
-    "test:http": "docker compose -f ./ipc-http/docker-compose.yml up",
-    "test:http:server": "node --max-old-space-size=1536 ./ipc-http/server.js",
-    "test:http:main": "THE_OPERATION_ENDPOINT_HOSTNAME=localhost THE_OPERATION_ENDPOINT_PORT=3000 node --max-old-space-size=1536 ./ipc-http/main.js",
+    "test:benchmark": "docker compose -f ./ipc-methods/benchmark/docker-compose.yml up",
+    "test:benchmark:raw": "npm run tsc && node --max-old-space-size=1536 ./ipc-methods/benchmark/main.js",
+    "test:http": "docker compose -f ./ipc-methods/http/docker-compose.yml up",
+    "test:http:server": "node --max-old-space-size=1536 ./ipc-methods/http/server.js",
+    "test:http:main": "THE_OPERATION_ENDPOINT_HOSTNAME=localhost THE_OPERATION_ENDPOINT_PORT=3000 node --max-old-space-size=1536 ./ipc-methods/http/main.js",
     "test:http:raw": "npm run tsc && concurrently \"npm run test:http:server\" \"npm run test:http:main\"",
-    "test:http-express-axios": "docker compose -f ./ipc-http-express-axios/docker-compose.yml up",
-    "test:http-express-axios:server": "node --max-old-space-size=1536 ./ipc-http-express-axios/server.js",
-    "test:http-express-axios:main": "THE_OPERATION_ENDPOINT_HOSTNAME=localhost THE_OPERATION_ENDPOINT_PORT=3000 node --max-old-space-size=1536 ./ipc-http-express-axios/main.js",
+    "test:http-express-axios": "docker compose -f ./ipc-methods/http-express-axios/docker-compose.yml up",
+    "test:http-express-axios:server": "node --max-old-space-size=1536 ./ipc-methods/http-express-axios/server.js",
+    "test:http-express-axios:main": "THE_OPERATION_ENDPOINT_HOSTNAME=localhost THE_OPERATION_ENDPOINT_PORT=3000 node --max-old-space-size=1536 ./ipc-methods/http-express-axios/main.js",
     "test:http-express-axios:raw": "npm run tsc && concurrently \"npm run test:http-express-axios:server\" \"npm run test:http-express-axios:main\""
   },
   "repository": {
